breakpoints and stepping and running oh my!
With this commit, enough of the protocol is implemented for the gdbstub
to be considered _useful_!

The one major feature that's broken (and that I've been trying to fix)
is watchpoints. GDB seems to recieve the watchpoint stop request, but
then immediately sends another vCont packet! Why?? I'll try to figure it
out.

...

The plan for the next two commits is two-fold:
1) fix watchpoints
2) re-evaluate and re-architect the API, especially Target::Usize

If I can get those two features working, I'd be super happy, since at
the very least, it would cover my own use-case.