rewrite internals to remove alloc dependency!!!
This is something I've wanted to do for a while now, so I sat down and
did it! As is usually the case with revisiting old code, you have to
wonder why I didn't write it this was in the first place. Ah well.

* * *

I've got at least 3 more features I want to implement before publishing
a new version:

1) Making a new `Arch` trait which encodes CPU register mappings,
pointer size, etc...
  - This aught to make the `read/write_registers` methods a whole lot
    easier to implement, since it won't rely on raw u8-streams who's
    ordering depends on an obscure XML document within the GDB repo.
2) Multiprocess extensions
  - This is a selfish addition, since I'll need to debug multiple cores
    in `clicky`
3) Support `monitor` commands (#9)
  - I've realized it would be nice to be able to query the state of my
    emulator while stepping though the code (e.g: debug-print structs),
    and `monitor` commands seem like a clean way to integrate the
    functionality