add `monitor` command handler + rework pkt parsing
This commit closes #9, and allows targets to handle arbitrary commands
sent via the `monitor` GDB command!

Mid-way though adding the monitor command support, I had a revelation:
why am I using an iterator to lazily parse hex-encoded bytes, when I
could just decode the hex data _in place_ of the packet buffer, and
return a slice directly to the decoded &[u8]??

Well, that idea was simple in practice, but surprisingly difficult to
implement due to my Incredible Genius of only passing `&str` to each
packet, instead of a raw `&mut [u8]`. It took a lot of futzing around,
but I reworked everything so that each packet can access the raw
undlerlying &mut [u8] bufer, and _optionally_ cast it to a `&str`.

I intend to use `lexical_core` at some point to parse numbers directly
from the &[u8] (instead of casting to a string), which should result in
a leaner binary, and some cleaner code (rust's built-in str brings in a
lot of UTF-8 machinery, which isn't required _at all_ in the GDB Remote
Serial Protocol, which is strictly ASCII (7-bit clean ASCII at that
lol).