add multithreading support!
It was a _big_ push, but here we are: `gdbstub` finally supports
debugging multithreaded targets!

I worked hard to keep the single-threaded use-case as simple as
possible, and I _think_ i've mostly pulled it off. The only ugly bit is
the use of the `SINGLE_THREAD_TID` constant, but IDK how else to do
it... Maybe some more experimentation will result in a nice solution...

With this feature finally implemented, I think I'll take some time to do
one more cleanup + documentation pass over the codebase, and publish
version 0.2.0 of the library to crates.io.

I'm excited to re-integrate `gdbstub` into `clicky`, and get cracking on
tackling more of the platform's multi-core functionality!

* * *

Some misc improvements included in this commit include:
- switching the `write_register` method from an iterator to a slice
- removing a bunch of num-cast unwraps