make `Target` object-safe + Arch trait tweak
If only there was a way to get the concrete type of the various closurs
I pass from gdbstub... It'd avoid the overhead of dynamic dispatch,
since those closures never change. Ah well!

The Arch trait tweak was switching the u8 iterator to a slice (which I
forgot to do when doing the first decode_hex_buf pass)