add GdbStub builder + support resuming sessions
I kinda had a half-assed builder pattern going on before this commit,
whereby the GdbStub itself would be a builder, and the `run` method
acted as a build + execute for the real `GdbStubImpl`. This commit fixes
it to be less jank.

Moreover, GdbStub's run method now uses `&mut self`, which means
sessions can be resumed in the case of a recoverable error! I'm going to
use this in my `clicky` project to support opening post-morem GDB
sessions...

I've also tweaked how some of the error types work. It leads to somewhat
longer typenames, but since they don't leak out into the public
interface, it's fine.