first-pass at trait-based optional feature support
Honestly, I'm pretty happy with how this is looking.

This current approach is reminiscent of a C-style
struct-of-function-pointers implementation, but with Traits instead. I
thought of this approach a while back, but never got around to
implementing it. Glad I found the time though, since it's really cool!

While the API is somewhat "unorthadox", the final implementations look
quite clean, and are easier to follow than having the single monolithic
`Target` method as in previous versions. Plus, I was able to remove the
gnarly `OptResult` type, which I always felt was really dirty!

Similarly, the internal implementation could be cleaned up quite a bit
as well, since checking for a supported feature doesn't actually requite
_invoking_ an associated method. Just check if the ops "pointer" is None
or not!

* * *

There are a couple more things I'd like to experiment with:

- is there some way that GdbStub's internal implementation can be
broken-up similarly to the public API? i.e: is there some way to move
the packet parsing + handler code for, say, `handle_monitor_cmd` into
it's own little module, which could then be dead-code-eliminated if a
target doesn't end up using that feature?
- more generally, what's the overhead of this abuse-of-dyn-traits
approach? I vaguely recall playing around with this idea in godbolt, and
being surprised by how effective the compiler was able to inline
everything, though I should probably run some more proper experiments...
  - Heck, that would probably be an interesting topic for a blog post!