Add dead code elimination hints
If a target doesn't implement the `MonitorCmd` extension, then it seems
like a shame to have it include a bunch of unused code related to
parsing + handling `qRcmd` packets, right?

Well, with some clever match guard hinting + the power of Inline Dyn
Extension Traits, it's possible to trick the Rust compiler into
optimizing unused packets out of the final binary!

Now, i'm no compiler engineer, so I can't say whether or not this
optimization is "guaranteed" to occur or not. That said, given how
simple the __protocol_hint method is, and how it can monomorphize to a
static `true / false` boolean, I strongly suspect that it will almost
always perform the optimization.

To assist in implementing + debugging this feature, I've included a new
internal `__dead_code_marker!` macro (and associated feature), which are
used by `script/test_dead_code_elim.sh` to check whether or not a
certain packet is optimized out in the example binaries. See the inline
documentation for how that script works...