finally clean up multiprocess extension support
Ughh, it took a long while, but I've finally come up with an API for
multiprocess support that I'm happy with, _and_ fixed some icky hacks
and bad implementation choices in the core gdbstub implementation.

Some interesting things that come out of this (chonker of a) commit:

- Turns out the `qC` packet is totally useless, so I've gone ahead and
removed it entirely. The less code the better.
- I've finally fixed the incredibly confusing naming scheme surrounding
the thread-id structures. The new names should be much more sane (and
have less overlap)
- I've massively improved the `resume` method for multithreaded targets.
The key insight is that returning a tuple of `(Tid, StopReason)` wasn't
exactly correct, as there are certain stop reasons that are _not_
associated with any Tid (e.g: GDB interrupt), and it effectively forced
the target to "flip a coin" as to which Tid it reported. This old system
wreaked havoc when dealing with breakpoints in multithreaded systems,
but with the new `ThreadStopReason` API, i've managed to clean up the
library internals to keep the GDB nice and happy, and removed the burden
of making arbitrary decisions from the Targer implementation.
- vCont works now! I'm a dummy, and never actually realized that my
vCont implementation was totally busted. I've gone ahead and fixed it,
so it should be working now.