Forward raw GDB register ID instead of unit
Upon further thought, I've realized that using the unit type as the
"unimplemented" RegId type is actually quite detrimental to the end-user
experience. It forces the user to fork `gdbstub` if they want to
implement the `read/write_register` methods for a built in architecure,
which could drive users away...

By passing through the raw GDB identifier, it gives users a chance to
"rough it out" on their own, while urging them to upstream their
implementations as well.