Make SwBreakpoint optional + tweak Target docs
I stumbled across the following line in the GDB RSP docs:

"At a minimum, a stub is required to support the ‘g’ and ‘G’ commands
for register access, and the ‘m’ and ‘M’ commands for memory access.
Stubs that only control single-threaded targets can implement run
control with the ‘c’ (continue), and ‘s’ (step) commands. Stubs that
support multi-threading targets should support the ‘vCont’ command.
All other commands are optional."

While I won't go so far as making `vCont` (i.e: resume) optional, this
does imply that breakpoints are not strictly required per the core
protocol.

Similarly, if I were to follow the letter of the docs, then the
`list_active_threads` method for multithreaded targets should also be
optional, but it's used to get a valid initial value for
`current_mem_tid` in the GdbStub implementation, so I'm keeping it
required for now.

* * *

Additionally, I've tweaked some of the docs + type aliases related to
the Target trait to make it slightly more obvious how IDETs work.