Add `write_all` and `flush` to `Connection`
While I'm not actually using these methods in the code just yet, I have
a feeling that it might be possible to speed up some of the outbound
communication by using these new methods.

This doesn't complicate the API too much, as `write_all` comes with a
default implementation, and `flush` can usually be implemented in only a
line or two (at least when the backing stream implements
std::io::Write).