lift `RegId` associated-type to `Arch` trait
While it would be _great_ if each RegId enum had a 1:1 correpspondance
to an associated Registers struct, the existance of target features
complicates this significantly.

Instead, it's probably fine to have a single `RegId` enum that
encompasses all of an `arch`'s registers (including features), and have
the user return `Ok(false)` for those registers they aren't using.

A happy side-effect of this change is cleaning up some nested
fully-qualified-type syntax.