add perliminary Extended Mode support
This commit also introduces some new error handling patterns, which I
intend to port over to the rest of the `Target` interface in the next
couple commits.

Namely, I've realized that the idealist approach of "just use ?" isn't
particularly feasable, as there are plenty of non-fatal errors that
users need to be able to report as well. The current approach has been
to use something janky, like returning `Ok(bool)` to denote
success/failure, but that's not really ideal.

This new approach is closer to the old `0.2.0` approach, albeit with
less "magic". Plus, given that the error cases are "fatal", having to
wrap errors in a `TargetError::Fatal` should make it easier to find
failure paths.