(protocol) add extension-based packet parsing
No more `__protocol_hint`! This new approach makes a lot more sense, and
makes it easy to see at-a-glance which packets are related to which
protocol extension.

As it happens, this actually _increases_ the no_std example from the
previous commit, but only by a tiny ammount. I haven't looked into it
whatsoever, but my guess might be that the nested-enums could be making
the optimizer's job harder? Just a guess though.

All-in-all, these last two commit have made the IDET appraoch a lot
less "hacky" internally, which should make the codebase easier to
maintain.