consolidate OK responses
While I did find the old approach of writing "OK" manually more
"aesthetically" pleasing, it made it easy to accidentally forget to
include it / omit it, and did result in some binary bloat.

These changes require each packet handler to explicity specify whether
or not it's done handling the response, or if an "OK" still needs to be
sent.