release 0.4.0
There's always more to add, but I've been sitting on these changes for
far too long, and it's about time to push out a proper release.

If I had to guess what the next breaking change is going to be, it'd
probably have to do with adding proper multiprocess support. I've been
experiementing with it on another branch, and while there's certainly
the "quick and dirty" approach of adding a new `base::MultiProcess`
trait + adjusting all the internal match arms to use it, I feel like
there might be some way to treat SingleThread/MultiThread as simple
"facades" over a One True underlying multiprocess-based
implementation...

Well, that's not here nor there...

Enjoy `0.4.0`!