fix vCont with signal parsing
I've never actually hit this code path until now, so this bug slipped
under the radar. It doesn't actually matter though, since resuming with
a singnal isn't supported in `gdbstub`.

I should really add support for stopping with signal, as there's no real
reason why it's not in yet... It'll be a breaking change though, which
sucks.