almost complete GDB agent + conditional breakpoint support
wowie what a doozy. getting this working nicely required a lot of
through and refactoring, though as it stands I'm pretty happy with how
the API is shaping up.

One notable missing piece as of this commit is that the gdbstub doesn't
actually evaluate any conditions/commands after a breakpoint it hit.
See, I realized that the Sw/HwBreak stop reasons don't actually include
the breakpoint address (I mean, why would they), but I need the
breakpoint address to ask the target for a list of BytecodeIds to
execute.

Oof.

I think I'm going to have to do something cheecky and modify the
`Registers` trait to include a `get_pc` method, which is kinda stinky,
but I don't see any better way... That way, I can simply call
`get_registers` after a breakpoint is hit and extract the PC.

This seems cleaner than the alternative, which would be to package the
PC value alongside the Sw/HwBreak stop reason. That would be a very
annoying breaking change, as opposed to the `Registers` trait tweak,
which is technically also a breaking change, but given that most folks
are using the built-in register defenitions, it shouldn't break too many
downtream users.