split single register access (P,p) into separate IDET
These methods were part of the base-ops because once upon a time,
`gdstbub` tried to be an "opinionated" library wrt. which extensions all
targets should implement. i.e: the end goal was to stabilize the RegId
types, and remove the defaul implementation of `read/write_register`,
forcing all targets to implement them.

Over time, `gdbstub` has shifted away from holding these sorts of
opinions, and instead strives to be as _modular_ as possible, only
requiring those features that _must_ be implemented as-per the GDB RSP
documentation.