rework state machine API
pretty much just cribbed all the ideas from
https://hoverbear.org/blog/rust-state-machine-pattern/

it did bump up the code size somewhat, which is unfortunate...
I wonder if the compiler isn't smart enough to realize that state
transitions are noops, and is inserting useless memcpys? I'll have to
dig into that some more...

nonetheless, this new API is a _lot_ tighter, and doesn't suffer from
the "spoofed state token" issue that the last API had. as such, it is
100% impossible to misuse the API and perform an operation out of
sequence.