reorganize public interface of `gdbstub`
over time, as more types and traits have been added to `gdbstub`, it's
become harder to judge which types deserve to live at the "top level" of
the crate vs. under submodules.

to simplify this cognitive load, I've tweaked the public API to reflect
the internal organization of the project a bit more.

this isn't a purely selfish change though - IMO this new organization
makes it easier to tell what types/traits are associated with what bits
of `gdbstub` functionlity.

With this change, the top-level `gdbstub::` namespace no longer contains
any concrete types/traits, and simply exposes a collection of modules.