remove vestigial "recoverable fatal error" handler
this bit of code has been in `gdbstub` for as long as I can remember,
and taking a look at it now, it's obvious that this code is totally
bogus.

it made a _bit_ of sense back when `resume` could return a fatal error,
in which case, it made sense to return a stop reason, thereby
"recovering" from the fatal error (at least from the GDB client's
perspective).

...but aside from that one use case, it's totally bogus. if you return a
fatal error from any other method aside from `resume`, the GDB client
would die with a packet-parsing error, since it would get a stop reply
packet entirely out of the blue!