Report thread-id in response to `?` packet
When more than one thread is active the following warning is visible upon connecting:

```
warning: multi-threaded target stopped without sending a thread-id, using first non-exited thread
```

This is because gdbstub sends a stop packet of `S05`, which conveys no thread-id.  This fix changes the response to `T05thread:<tid>;`, which eliminates the warning.

Signed-off-by: Ryan Fairfax <ryan@thefaxman.net>