core_impl: fix support for no threads
Sometimes, a target has no threads. This may be the case when a process
is not yet selected, or when all threads have terminated in a specific
process.

Support `None` as a return from `get_sane_any_tid()` and return an error
to GDB. This addresses a `NoActiveThreads` error, and closes #122.

Signed-off-by: Sean Cross <sean@xobs.io>