vAttach: get vAttach working with multi-process (#129)
* vAttach: get vAttach working with multi-process

Get vAttach working to the point where it is now possible to switch
processes by running `attach [PID]` in GDB.

Signed-off-by: Sean Cross <sean@xobs.io>

* base: move `report_stop_reason` into its own function

Create a central place for reporting the stop reason. This will be used
when GDB issues a `QuestionMark` packet.

Signed-off-by: Sean Cross <sean@xobs.io>

* extended_mode: use `report_stop_reason` for vAttach

When attaching to a process, report the stop reason using the new
`report_stop_reason()` function. This enables attaching to processes
without hardcoding the `S00` response.

Signed-off-by: Sean Cross <sean@xobs.io>

* base: qC: return self.current_mem_tid if it's set

If self.current_mem_tid is set to a value other than the default
sentinal value, return the current thread ID.

Otherwise, look through the list of active threads and return the first
active one.

Signed-off-by: Sean Cross <sean@xobs.io>

---------

Signed-off-by: Sean Cross <sean@xobs.io>