fix addr validity check in armv4t example
the emulator is dead simple, and has RAM covering the whole 32 bit
address space. The old implementation assumed that the link script
totally encapsulated available RAM, when that isn't actually the case.

The more "proper" solution would probably be to fix up the emulator and
link script to be more consistent with one another... but this seems
like a reasonable enough fix as well.