Flash operations support (#172)
This PR adds support for the flash commands as described in
(https://sourceware.org/gdb/current/onlinedocs/gdb.html/Packets.html#Packets).

I had to change the memory map in the armv4t example providing a real
memory map, otherwise GDB would not load a real .elf file. It doesn't
have to be a real memory map, but you need at least a region marked a
flash containing the address of the elf sections to be loaded. (eg:
.vector_table, .text etc..)