error: support core::error::Error types
Support reporting errors that conform to core::error::Error. This allows for
calling e.source(), which may provide more information.

As an example, using the `probe-rs` gdbserver. Before:

During the execution of GDB an error was encountered:
Target threw a fatal error: A RISC-V specific error occurred.
 (No further information is available)

After:

During the execution of GDB an error was encountered:
Target threw a fatal error: A RISC-V specific error occurred.
   ... caused by: Error occurred during execution of an abstract command: Exception

This is gated by "std", so should not pose any API changes to deeply
embedded targets.